pyproject.toml
[tool.poetry]
name = "free-simple-media-organizer"
version = "1.00.00"
description = "Darmowy prosty ogranizer plików multimedialnych napisany w Python"
authors = ["Mateusz Mocydlarz <mateusz.mocydlarz@bitmm.eu>"]
license = "MIT"
readme = "README.md"
packages = [{include = "free_simple_media_organizer"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.19"
alembic = "^1.11.2"
python-magic-bin = "^0.4.14"
python-magic = "^0.4.27"
filetype = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
